requirements.txt
# Django and REST framework
Django>=3.2.0
djangorestframework>=3.12.0

# Redis integration
redis>=4.0.0
django-redis>=5.0.0  # Django Redis integration

# HTTP requests
requests>=2.28.0  # For making HTTP requests

# JWT Authentication (needed for login_user view's RefreshToken)
djangorestframework-simplejwt>=5.0.0

# Django Utilities
django-cors-headers>=3.10.0  # Recommended for API development

# Time zone support
pytz>=2021.1

# For async support (used in ASGI)
asgiref>=3.4.0

# For database connection and operations
sqlparse>=0.4.0

# Type checking support
typing-extensions>=4.0.0

# For environment variables (recommended for settings management)
python-dotenv>=0.19.0
